fix: F2 프리셋 화이트리스트 + F3 재클론 시 상한 재적용 (Codex #599 P2)
P2-1(tts.ts): 기본/무료 보이스 프리셋 게이트를 love-only 블랙리스트에서
FREE_BUCKET_CATEGORIES(날씨+약) 화이트리스트로 교체 — morning/health/exercise 등
다른 프리셋 카테고리가 시스템 보이스로 새어 합성되던 우회 차단.

P2-2(voice-slots.ts): eviction 로직(MAX_PROVIDER_CLONE_VOICES + evictLruClonesIfOverCap)을
lib로 추출해 /clone 과 F3 재클론(voice-recover)이 공유. 재클론도 enroll 직전 상한을
재적용해, 이미 상한이면 다른 LRU를 비운 뒤 복원(상한 초과 51 방지).

테스트: 시스템 보이스 프리셋 테스트 2종을 medication(허용 카테고리)으로 갱신. 1537 통과.
